+++ SchoolRush/Assets/Scripts/Data/PlayerDataSaveLoad.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 34ce90cc0201840eb87d1aec4b5d00b9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
