--- SchoolRush/Assets/Scripts/UI/PlayerData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e9a396545e48e4dd6ab4dc90a30c6fcf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
