--- SchoolRush/Assets/TextMesh Pro/Resources/Fonts & Materials/LiberationSans SDF.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8f586378b4e144a9851e7b34d9b748ee
-timeCreated: 1484171803
-licenseType: Pro
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
